LPS-48075 Move things around
