+++ modules/util/ip-geocoder/src/com/liferay/ip/geocoder/IPInfo.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.ip.geocoder.model;
+package com.liferay.ip.geocoder;
 
 import com.maxmind.geoip.Location;
